backend/requirements.txt
# Web Framework
fastapi>=0.110.0,<0.116.0
uvicorn[standard]>=0.27.0,<0.32.0

# Database
sqlalchemy[asyncio]>=2.0.27,<2.1.0
alembic>=1.13.1,<1.14.0
asyncpg>=0.29.0,<0.30.0
pgvector>=0.2.5,<0.3.0
greenlet>=3.0.0

# Data Validation
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.0.1  # passlib 1.7.4 is incompatible with bcrypt 4.1+
google-auth>=2.0.0,<3.0.0  # Google id_token verification
requests>=2.32.0,<3.0.0    # Required transport layer for google-auth

# File Upload
python-multipart>=0.0.9,<0.1.0

# AWS/Storage
boto3>=1.34.0,<2.0.0

# AI/ML
openai>=1.12.0,<2.0.0

# Cache
redis>=5.0.0,<6.0.0

# Background Tasks
celery>=5.3.0,<6.0.0

# HTTP Client
httpx>=0.27.0,<0.28.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
